Splitting up catalog.py in multiple files.
